avocado.virt.qemu.machine: Add screendump thread to VMs
Create a screendump thread on every VM. This is a departure
from previous efforts (virt-test), where the thread was
under control of an extraneous part of the code.

So at the QEMU process init stage, start the thread, when
the VM is shut down, stop it. This way, we have images
of what was going on in the VM during the test process.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>