avocado.virt.qemu.machine: Add screendump thread to VMs
Create a screendump thread on every VM (optional,
customizable through avocado.args.virt.screendump_thread_enable).
This is a departure from previous efforts (virt-test), where
the thread was not under class control. The frequence and
quality of the saved screenshots (.jpgs) can be controlled
through params as well.

So at the QEMU process init stage, start the thread, when
the VM is shut down, stop it. This way, we have a visual record
of what was going on in the VM during the test process.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>