avocado.virt.qemu.machine: Add uuid as a Machine() constructor param
Migration is a process by which VM state is transferred
to another instance of the virtualization system transparently,
and with virtually no downtime. This motivated us to model
the .migrate() method of the Machine() class preserving the
instantiated object (for all intents and purposes, it is
the 'same' virtual machine).

Therefore, it makes sense to have videos of a vm being migrated
as a single entity. Make uuid as one of the Machine class
constructor params, so that we have the option of making
migrated vms to keep the same uuid, and have them to behave
as a single entity for the purposes of vm video generation.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>