avocado.virt.qemu.machine: Encode videos upon VM object termination
As soon as the machine is powered off, let's create a
video from the screenshots generated by the screendump
thread. In order to not encode videos unnecessarily
during migration, pass a migrate param to encode videos
and power off/screendump thread end methods to skip
video encoding in such ocasions.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>